build: fix build failure due to many factors
This commit fixes failed build
https://circleci.com/gh/huy-nguyen/remark-github-plugin/7.

The reason is multifold:

1) In commit f8eb4781 when I first started using `node-fetch`, I forgot
to save it to `package.json`. Fetching still worked because `node-fetch`
v1.7.3 was still installed because `eslint-plugin-react` depended on it.
This version of `node-fetch` does not specify a `browser` field in its
`package.json`.

2) In commit a9d70ec2 (which caused the build failure), `node-fetch`
v2.1.2 was installed because `semantic-release` depends on it. This
version of `node-fetch` does have a `browser` field which points to a
build result that depends on the `window` global variable, which doesn't
exist in `node`.

3) `rollup` is accidentally set to treat the `pkg.browser` field as
having higher precedence than `pgk.module` when resolving packages. Thus
it reads from the `browser` field of `node-fetch`, which causes
`rollup`'s built file to fail in tests.

Solution:

1) explicitly install `node-fetch`.

2) disable `pkg.browser` in `rollup` config.