fix: correct query parameter in AJAX call
The correct query parameter to get a file from GitHub API is `ref` but
we've been using `branch` this whole time. This has worked so far
because all the tests have been performed on a file
(`https://github.com/huy-nguyen/squarify/blob/d7074c2/.babelrc`) that
has only one version in git history. Providing a `branch` parameter
instead of a `ref` parameter causes GitHub to assume no `ref` has been
specified and as such to return the latest version of that file on the
`master` branch, which coincidentally is the correct content of the
tested file (which ever only has one version).

This commit fixes that error and also updates all test files so that
they don't use the special case of files with only one version in git
history.