build: revert @babel-scoped packages to 7.0.0-beta.42
It turns out build
https://circleci.com/gh/huy-nguyen/remark-github-plugin/70 failed after
a dependency update PR (#11) by `greenkeeper` because of a breaking
change in `@babel` going from 7.0.0-beta.42 to 7.0.0-beta.46
(https://github.com/babel/babel/pull/7358). This breaking change in
`babel` changes the scope of `.babelrc.js` and causes files inside
`lodash-es` (which are written as untranspiled ES modules) to not be
transpiled to CommonJS. This in turn causes `jest` to fail.

Also revert the `--no-cache` CLI option for test command on CI because
it wasn't the cause of the problem.