ci: make greenkeeper's lock file update run before compilation
The reason is that if `greenkeeper-lockfile-update` is run after the
build command, the build output (which is what's tested by `npm run
test:ci`) wouldn't incorporate any new packages that greenkeeper
downloaded.  In that case, if the new dependency version that
greenkeeper downloaded actually breaks the tests, greenkeeper would open
an erroneous PR (to recommend that we upgrade to the breaking
dependency).